added shutting_down state
this is like restarting except it does not actually try to restart the server. this is meant only for bungeecord servers the shutdown at the end of the game. when using the default restarting state, a new game would start while the server was being shut down which obviously isn't ideal behavior